# MemoryMappedFile: don't check for file existence if there is no need