@EgorBo 2022-01-18
Improvement on win-x64 FullPGO: https://github.com/dotnet/perf-autofiling-issues/issues/2958